requirements.txt
torch
pandas
seaborn
bert-serving-client